refactor(app): presetswitcher refactoring
